+++ 3rdPartyDrivers/lego-sound.h
@@ -5,6 +5,10 @@
  * @{
  */
 
+/*
+ * $Id: lego-sound.h 133 2013-03-10 15:15:38Z xander $
+ */
+
 /** \file lego-sound.h
  * \brief SMUX driver for the Lego Sound sensor.
  *
@@ -16,7 +20,7 @@
  *
  * License: You may use this code as you wish, provided you give credit where its due.
  *
- * THIS CODE WILL ONLY WORK WITH ROBOTC VERSION 4.10 AND HIGHER
+ * THIS CODE WILL ONLY WORK WITH ROBOTC VERSION 3.59 AND HIGHER. 
 
  * \author Xander Soldaat (xander_at_botbench.com)
  * \date 25 November 2009
@@ -31,8 +35,8 @@
 #endif
 
 #ifdef __HTSMUX_SUPPORT__
-short SNDreadRaw(tMUXSensor muxsensor);
-short SNDreadNorm(tMUXSensor muxsensor);
+int SNDreadRaw(tMUXSensor muxsensor);
+int SNDreadNorm(tMUXSensor muxsensor);
 void SNDsetDBA(tMUXSensor muxsensor);
 void SNDsetDB(tMUXSensor muxsensor);
 
@@ -41,19 +45,21 @@ void SNDsetDB(tMUXSensor muxsensor);
  * @param muxsensor the SMUX sensor port number
  * @return the raw value of the Sound Sensor
  */
-short SNDreadRaw(tMUXSensor muxsensor) {
+int SNDreadRaw(tMUXSensor muxsensor) {
   return 1023 - HTSMUXreadAnalogue(muxsensor);
 }
+
 
 /**
  * Get the processed value from the sensor.
  * @param muxsensor the SMUX sensor port number
  * @return processed value of the sensor or -1 if no valid range has been specified.
  */
-short SNDreadNorm(tMUXSensor muxsensor) {
-  short _val = ((long)SNDreadRaw(muxsensor) * (long)100) / (long)1024;
+int SNDreadNorm(tMUXSensor muxsensor) {
+  int _val = ((long)SNDreadRaw(muxsensor) * (long)100) / (long)1024;
   return _val;
 }
+
 
 /**
  * Set the mode of the sensor to DBA, this is done
@@ -63,6 +69,7 @@ short SNDreadNorm(tMUXSensor muxsensor) {
 void SNDsetDBA(tMUXSensor muxsensor) {
   HTSMUXsetAnalogueInactive(muxsensor);
 }
+
 
 /**
  * Set the mode of the sensor to DB, this is done
@@ -74,5 +81,8 @@ void SNDsetDB(tMUXSensor muxsensor) {
 }
 #endif // __HTSMUX_SUPPORT__
 
+/*
+ * $Id: lego-sound.h 133 2013-03-10 15:15:38Z xander $
+ */
 /* @} */
 /* @} */
