+++ 3rdPartyDrivers/mindensors-servo.h
@@ -5,6 +5,10 @@
  * @{
  */
 
+/*
+ * $Id: mindensors-servo.h 133 2013-03-10 15:15:38Z xander $
+ */
+
 #ifndef __NXTSERVO_H__
 #define __NXTSERVO_H__
 /** \file mindensors-servo.h
@@ -20,7 +24,7 @@
  *
  * License: You may use this code as you wish, provided you give credit where its due.
  *
- * THIS CODE WILL ONLY WORK WITH ROBOTC VERSION 4.10 AND HIGHER
+ * THIS CODE WILL ONLY WORK WITH ROBOTC VERSION 3.59 AND HIGHER. 
 
  * \author Xander Soldaat (xander_at_botbench.com)
  * \date 30 September 2009
@@ -78,12 +82,13 @@
 <function prototypes>
 */
 bool NXTServoSetSpeed(tSensors link, ubyte servochan, ubyte speed, ubyte address);
-bool NXTServoSetPos(tSensors link, ubyte servochan, short position, ubyte speed, ubyte address = NXTSERVO_I2C_ADDR);
+bool NXTServoSetPos(tSensors link, ubyte servochan, int position, ubyte speed, ubyte address = NXTSERVO_I2C_ADDR);
 bool NXTServoQSetPos(tSensors link, ubyte servochan, ubyte position, byte speed, ubyte address = NXTSERVO_I2C_ADDR);
-short NXTServoReadVoltage(tSensors link, ubyte address = NXTSERVO_I2C_ADDR);
+int NXTServoReadVoltage(tSensors link, ubyte address = NXTSERVO_I2C_ADDR);
 
 tByteArray NXTSERVO_I2CRequest;         /*!< Array to hold I2C command data */
 tByteArray NXTSERVO_I2CReply;           /*!< Array to hold I2C reply data */
+
 
 /**
  * Set the speed register for the specified servo.  This is the amount to increase
@@ -106,6 +111,7 @@ bool NXTServoSetSpeed(tSensors link, ubyte servochan, ubyte speed, ubyte address
   return writeI2C(link, NXTSERVO_I2CRequest);
 }
 
+
 /**
  * Tell the servo to move to the specified position using the specified speed.
  * @param link the NXTServo port number
@@ -115,7 +121,7 @@ bool NXTServoSetSpeed(tSensors link, ubyte servochan, ubyte speed, ubyte address
  * @param address the I2C address to use, optional, defaults to 0xB0
  * @return true if no error occured, false if it did
  */
-bool NXTServoSetPos(tSensors link, ubyte servochan, short position, ubyte speed, ubyte address) {
+bool NXTServoSetPos(tSensors link, ubyte servochan, int position, ubyte speed, ubyte address) {
   memset(NXTSERVO_I2CRequest, 0, sizeof(tByteArray));
   if (!NXTServoSetSpeed(link, servochan, speed, address))
     return false;
@@ -132,6 +138,8 @@ bool NXTServoSetPos(tSensors link, ubyte servochan, short position, ubyte speed,
 
   return writeI2C(link, NXTSERVO_I2CRequest);
 }
+
+
 
 /**
  * Tell the servo to move to the specified position using the specified speed.
@@ -162,13 +170,14 @@ bool NXTServoQSetPos(tSensors link, ubyte servochan, ubyte position, byte speed,
   return writeI2C(link, NXTSERVO_I2CRequest);
 }
 
+
 /**
  * Get the voltage level of the battery pack.
  * @param link the NXTServo port number
  * @param address the I2C address to use, optional, defaults to 0xB0
  * @return the voltage of the battery pack or -1 if an error occurred
  */
-short NXTServoReadVoltage(tSensors link, ubyte address) {
+int NXTServoReadVoltage(tSensors link, ubyte address) {
   long mvs = 0;
 
   memset(NXTSERVO_I2CRequest, 0, sizeof(tByteArray));
@@ -186,6 +195,8 @@ short NXTServoReadVoltage(tSensors link, ubyte address) {
 }
 
 #endif // __NXTSERVO_H__
-
+/*
+ * $Id: mindensors-servo.h 133 2013-03-10 15:15:38Z xander $
+ */
 /* @} */
 /* @} */
