+++ 3rdPartyDrivers/timer.h
@@ -5,6 +5,10 @@
  * @{
  */
 
+/*
+ * $Id: timer.h 133 2013-03-10 15:15:38Z xander $
+ */
+
 #ifndef __TMR_H__
 #define __TMR_H__
 /** \file timer.h
@@ -18,7 +22,7 @@
  * before this driver is included.
  *
  * License: You may use this code as you wish, provided you give credit where its due.
- * THIS CODE WILL ONLY WORK WITH ROBOTC VERSION 4.10 AND HIGHER
+ * THIS CODE WILL ONLY WORK WITH ROBOTC VERSION 3.59 AND HIGHER. 
 
  *
  * Changelog:
@@ -38,6 +42,7 @@
 #define MAX_TIMERS 10  /*!< Maximum number of _timers */
 #endif
 
+
 /*!< Struct for timer data */
 typedef struct {
   long startTime;
@@ -47,11 +52,11 @@ typedef struct {
 typeTMR _timers[MAX_TIMERS]; /*!< Array to hold timer data */
 
 // Prototypes
-short TMRnewTimer();
-bool TMRisExpired(short timerIdx);
-void TMRreset(short timerIdx);
-void TMRreset(short timerIdx, long duration);
-void TMRsetup(short timerIdx, long duration);
+int TMRnewTimer();
+bool TMRisExpired(int timerIdx);
+void TMRreset(int timerIdx);
+void TMRreset(int timerIdx, long duration);
+void TMRsetup(int timerIdx, long duration);
 
 /**
  * Create a new timer.  It's an index to the next available timer in the array of
@@ -59,20 +64,21 @@ void TMRsetup(short timerIdx, long duration);
  * @return the first available slot in the timer object array or -1 if all slots
  *         have been used.  Increase the MAX_TIMERS variable in this case.
  */
-short TMRnewTimer() {
-  static short _tmrIdx = -1;
+int TMRnewTimer() {
+  static int _tmrIdx = -1;
   if (_tmrIdx < (MAX_TIMERS - 2))
     return ++_tmrIdx;
   else
     return -1;
 }
 
+
 /**
  * Check if the timer has expired.
  * @param timerIdx the timer to be checked.
  * @return true if the timer has expired, false if it hasn't.
  */
-bool TMRisExpired(short timerIdx) {
+bool TMRisExpired(int timerIdx) {
   hogCPU();
   if (_timers[timerIdx].startTime < 0) {
     return true;
@@ -82,46 +88,50 @@ bool TMRisExpired(short timerIdx) {
   releaseCPU();
 }
 
+
 /**
  * Reset the timer, will also mark "expired" flag as false.\n
  * This function will also check if the TMRtask is running and
  * start it up if this isn't the case.
  * @param timerIdx the timer to be checked.
  */
-void TMRreset(short timerIdx) {
+void TMRreset(int timerIdx) {
   hogCPU();
-  _timers[timerIdx].startTime = nPgmTime;
-  releaseCPU();
+	_timers[timerIdx].startTime = nPgmTime;
+	releaseCPU();
 }
+
 
 /**
  * Reset the timer, will also mark "expired" flag as false.
  * @param timerIdx the timer to be checked.
  * @param duration the amount of time the timer should run for before expiring.
  */
-void TMRreset(short timerIdx, long duration) {
+void TMRreset(int timerIdx, long duration) {
   hogCPU();
   _timers[timerIdx].duration = duration;
-  _timers[timerIdx].startTime = nPgmTime;
-  releaseCPU();
+	_timers[timerIdx].startTime = nPgmTime;
+	releaseCPU();
 }
+
 
 /**
  * Cause the timer to expire.
  * @param timerIdx the timer to be expired.
  */
-void TMRexpire(short timerIdx) {
+void TMRexpire(int timerIdx) {
   hogCPU();
   _timers[timerIdx].startTime = -1;
   releaseCPU();
 }
+
 
 /**
  * Configure the duration of the timer.
  * @param timerIdx the timer to be checked.
  * @param duration the amount of time the timer should run for before expiring.
  */
-void TMRsetup(short timerIdx, long duration) {
+void TMRsetup(int timerIdx, long duration) {
   hogCPU();
   _timers[timerIdx].duration = duration;
   releaseCPU();
@@ -129,5 +139,8 @@ void TMRsetup(short timerIdx, long duration) {
 
 #endif // __TMR_H__
 
+/*
+ * $Id: timer.h 133 2013-03-10 15:15:38Z xander $
+ */
 /* @} */
 /* @} */
